Fix image scaling
